+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 ```
 
+## [Unreleased] - YYYY-MM-DD
+### Added
+- Added defaults for optional value arguments.
+### Changed
+### Deprecated
+### Fixed
+### Security
+
+
 ## [0.3.2] - 2024-02-09
 ### Fixed
 - Fixed bash argument passing to allow quoted strings with spaces as arguments